RA-20 ViewModel Recipe
по RA-18 Порядок организации методов: вопрос по приватным методам, которые используются в других приватных методах внутри класса, куда их пихать, в порядке вызовов после приватных методжов первого уровня? =)

RA-20
1. Как лучше назвать пакет для "Список рецептов" recipeList у меня сейчас.
2. Экстеншен и BigDecimal отправить в пакет UI?
3. data class RecipeState если я правильно понимаю, оно должно быть всё var и списки тоже, так как сами списке меняться не будут, а будут использоваться полностью новые
4. Нужно будет ещё пару раз перечитать теорию )
