+++ app/src/main/java/com/example/cobarecipesapp/ui/MainActivity.kt
@@ -1,10 +1,13 @@
-package com.example.cobarecipesapp
+package com.example.cobarecipesapp.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.commit
 import androidx.fragment.app.replace
+import com.example.cobarecipesapp.R
 import com.example.cobarecipesapp.databinding.ActivityMainBinding
+import com.example.cobarecipesapp.ui.categories.CategoriesListFragment
+import com.example.cobarecipesapp.ui.recipes.favorites.FavoritesFragment
 
 class MainActivity : AppCompatActivity() {
