+++ app/src/main/java/com/example/cobarecipesapp/ui/categories/CategoriesListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.cobarecipesapp
+package com.example.cobarecipesapp.ui.categories
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,7 +8,10 @@ import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import androidx.fragment.app.replace
+import com.example.cobarecipesapp.R
+import com.example.cobarecipesapp.data.STUB
 import com.example.cobarecipesapp.databinding.FragmentListCategoriesBinding
+import com.example.cobarecipesapp.ui.recipes.recipeList.RecipesListFragment
 
 class CategoriesListFragment : Fragment(R.layout.fragment_list_categories) {
