+++ app/src/main/java/com/example/cobarecipesapp/ui/recipes/Extensions.kt
@@ -1,4 +1,6 @@
-package com.example.cobarecipesapp
+package com.example.cobarecipesapp.ui.recipes
+
+import com.example.cobarecipesapp.ui.recipes.recipeList.RecipesListAdapter
 
 fun RecipesListAdapter.setOnItemClick(listener: (Int) -> Unit) {
     setOnItemClickListener(object : RecipesListAdapter.OnItemClickListener {
