+++ app/src/main/java/com/example/cobarecipesapp/ui/recipes/recipe/RecipeViewModel.kt
@@ -1,0 +1,14 @@
+package com.example.cobarecipesapp.ui.recipes.recipe
+
+import androidx.lifecycle.ViewModel
+import com.example.cobarecipesapp.model.Ingredient
+
+
+data class RecipeState(
+    var title: String? = null,
+    var ingredients: List<Ingredient> = emptyList(),
+    var method: List<String> = emptyList(),
+    var imageUrl: String? = null,
+)
+
+class RecipeViewModel : ViewModel()
